refactor #10 - 커스텀 훅 분리
